requirements.txt
cycler==0.10.0
ExifRead==2.1.2
imageio==2.4.1
kiwisolver==1.0.1
matplotlib==3.0.0
numpy==1.15.3
opencv-python==3.4.3.18
pandas==0.24.1
Pillow==5.3.0
pyparsing==2.2.2
python-dateutil==2.7.3
pytz==2018.9
scipy==1.1.0
simpy==3.0.11
six==1.11.0
